+++ include/mckl/core/state_matrix.hpp
@@ -108,11 +108,14 @@ class StateMatrixBase : private internal::StateMatrixDim<Dim>
         }
 
         /// \brief `this->particle().state()(this->i(), j)`
-        value_type &operator()(size_type j) const
+        value_type &operator[](size_type j) const
         {
             return this->particle().state()(
                 static_cast<size_type>(this->i()), j);
         }
+
+        /// \brief `this->particle().state()(this->i(), j)`
+        value_type &operator()(size_type j) const { return operator[](j); }
 
         /// \brief `this->particle().state().at(this->i(), j)`
         value_type &at(size_type j) const
@@ -120,7 +123,7 @@ class StateMatrixBase : private internal::StateMatrixDim<Dim>
             runtime_assert(
                 j < dim(), "**StateMatrix::at** index out of range");
 
-            return operator()(j);
+            return operator[](j);
         }
     }; // class particle_index_type